+++ Components/Component.php
@@ -58,7 +58,7 @@ abstract class Component implements UriComponentInterface, Conditionable
     final protected static function filterUri(WhatWgUrl|Rfc3986Uri|Stringable|string $uri): WhatWgUrl|Rfc3986Uri|UriInterface|Psr7UriInterface
     {
         return match (true) {
-            $uri instanceof Modifier => $uri->uri(),
+            $uri instanceof Modifier => $uri->unwrap(),
             $uri instanceof Psr7UriInterface, $uri instanceof UriInterface => $uri,
             $uri instanceof WhatWgUrl => Uri::new($uri->toAsciiString()),
             default => Uri::new($uri),